Revert "refactor: TimePicker to ts"
This reverts commit 85070a5f40b9ecdf76db2d22ef27ef57abaafa81.
